+++ angular.json
@@ -1,0 +1,41 @@
+{
+  "$schema": "./node_modules/@angular/cli/lib/config/schema.json",
+  "version": 1,
+  "newProjectRoot": "projects",
+  "projects": {
+    "ngx-virtual-scroller": {
+      "projectType": "library",
+      "root": "projects/ngx-virtual-scroller",
+      "sourceRoot": "projects/ngx-virtual-scroller/src",
+      "prefix": "lib",
+      "architect": {
+        "build": {
+          "builder": "@angular-devkit/build-angular:ng-packagr",
+          "options": {
+            "project": "projects/ngx-virtual-scroller/ng-package.json"
+          },
+          "configurations": {
+            "production": {
+              "tsConfig": "projects/ngx-virtual-scroller/tsconfig.lib.prod.json"
+            },
+            "development": {
+              "tsConfig": "projects/ngx-virtual-scroller/tsconfig.lib.json"
+            }
+          },
+          "defaultConfiguration": "production"
+        },
+        "test": {
+          "builder": "@angular-devkit/build-angular:karma",
+          "options": {
+            "main": "projects/ngx-virtual-scroller/src/test.ts",
+            "tsConfig": "projects/ngx-virtual-scroller/tsconfig.spec.json",
+            "karmaConfig": "projects/ngx-virtual-scroller/karma.conf.js"
+          }
+        }
+      }
+    }
+  },
+  "cli": {
+    "analytics": false
+  }
+}
