+++ package.json
@@ -1,14 +1,15 @@
 {
-  "$schema": "./node_modules/ng-packagr/package.schema.json",
   "name": "@iharbeck/ngx-virtual-scroller",
-  "version": "13.0.4",
-  "description": "Angular module for virtual -infinite- list. Supports horizontal/vertical, variable heights, & multi-column",
-  "main": "dist/virtual-scroller.js",
-  "types": "dist/virtual-scroller.d.ts",
-  "sideEffects": true,
+  "version": "14.0.1",
   "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test",
+
     "cleanup": "rimraf dist",
-    "build": "ng-packagr -p package.json",
+    "___build": "ng-packagr -p package.json",
     "build:docs": "rimraf demo/node_modules/ngx-virtual-scroller && cd dist && copyfiles **/* ../demo/node_modules/ngx-virtual-scroller && cd .. && rimraf docs && mkdir docs && cd demo && npm run build && cd dist/demo && copyfiles **/* ../../../docs && cd ../../.. && replace \"<base href=\\\"/\\\">\" \"<base href=\\\"/ngx-virtual-scroller/\\\">\" docs/index.html",
     "build:all": "npm run build && npm run build:docs",
     "build:all2": "npm run build",
@@ -20,79 +21,34 @@
     "prod-publish": "npm run build:all2 && npm publish --access=public ./dist",
     "publish-test": "npm pack ./dist"
   },
-  "ngPackage": {
-    "lib": {
-      "entryFile": "src/virtual-scroller.ts"
-    },
-    "allowedNonPeerDependencies": [
-      "@tweenjs/tween.js",
-      "@types/tween.js",
-      "@angular-devkit/build-angular"
-    ]
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/iharbeck/ngx-virtual-scroller.git"
-  },
-  "keywords": [
-    "virtual scroll",
-    "scroll",
-    "ngx",
-    "angular",
-    "angular 13"
-  ],
-  "contributors": [
-    {
-      "name": "Rinto Jose"
-    },
-    {
-      "name": "Devin Garner"
-    },
-    {
-      "name": "Pavel Kukushkin"
-    },
-    {
-      "name" : "Ingo Harbeck"
-    }
-  ],
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/iharbeck/ngx-virtual-scroller/issues"
-  },
-  "homepage": "https://github.com/iharbeck/ngx-virtual-scroller#readme",
-  "peerDependencies": {
-    "@angular/core": "^13.3.6"
-  },
+  "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "^13.3.5",
-    "@tweenjs/tween.js": "^18.6.4",
-    "@types/tween.js": "^18.6.1"
+    "@angular/animations": "^14.0.4",
+    "@angular/common": "^14.0.4",
+    "@angular/compiler": "^14.0.4",
+    "@angular/core": "^14.0.4",
+    "@angular/forms": "^14.0.4",
+    "@angular/platform-browser": "^14.0.4",
+    "@angular/platform-browser-dynamic": "^14.0.4",
+    "@angular/router": "^14.0.4",
+    "rxjs": "~7.4.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.11.4",
+    "@tweenjs/tween.js": "^18.6.4"
   },
   "devDependencies": {
-    "@angular/cli": "^13.3.5",
-    "@angular/common": "^13.3.6",
-    "@angular/compiler": "^13.3.6",
-    "@angular/compiler-cli": "^13.3.6",
-    "@angular/core": "^13.3.6",
-    "@angular/forms": "^13.3.6",
-    "@angular/language-service": "^13.3.6",
-    "@angular/platform-browser": "^13.3.6",
-    "@angular/platform-browser-dynamic": "^13.3.6",
-    "@angular/router": "^13.3.6",
-    "@tweenjs/tween.js": "^18.6.4",
-    "@types/node": "^16.11.33",
-    "codelyzer": "^6.0.2",
-    "copyfiles": "^2.4.1",
-    "core-js": "^3.22.4",
-    "ng-packagr": "^13.3.1",
-    "npm-check": "^5.9.2",
-    "replace": "^1.2.1",
-    "replace-x": "^1.7.2",
-    "rimraf": "^3.0.2",
-    "rxjs": "^7.5.5",
-    "tslib": "^2.3.1",
-    "tslint": "^6.1.3",
-    "typescript": "^4.4.4",
-    "zone.js": "^0.11.4"
+    "@angular-devkit/build-angular": "^14.0.4",
+    "@angular/cli": "^14.0.4",
+    "@angular/compiler-cli": "^14.0.4",
+    "@types/jasmine": "~3.10.0",
+    "@types/node": "^12.11.1",
+    "jasmine-core": "~3.10.0",
+    "karma": "~6.3.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "~2.0.3",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "~1.7.0",
+    "ng-packagr": "^14.0.2",
+    "typescript": "~4.7.4"
   }
 }
