+++ projects/ngx-virtual-scroller/README.md
@@ -1,0 +1,24 @@
+# NgxVirtualScroller
+
+This library was generated with [Angular CLI](https://github.com/angular/angular-cli) version 13.0.0.
+
+## Code scaffolding
+
+Run `ng generate component component-name --project ngx-virtual-scroller` to generate a new component. You can also use `ng generate directive|pipe|service|class|guard|interface|enum|module --project ngx-virtual-scroller`.
+> Note: Don't forget to add `--project ngx-virtual-scroller` or else it will be added to the default project in your `angular.json` file. 
+
+## Build
+
+Run `ng build ngx-virtual-scroller` to build the project. The build artifacts will be stored in the `dist/` directory.
+
+## Publishing
+
+After building your library with `ng build ngx-virtual-scroller`, go to the dist folder `cd dist/ngx-virtual-scroller` and run `npm publish`.
+
+## Running unit tests
+
+Run `ng test ngx-virtual-scroller` to execute the unit tests via [Karma](https://karma-runner.github.io).
+
+## Further help
+
+To get more help on the Angular CLI use `ng help` or go check out the [Angular CLI Overview and Command Reference](https://angular.io/cli) page.
