+++ projects/ngx-virtual-scroller/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "iharbeck-ngx-virtual-scroller",
+  "version": "0.0.1",
+  "peerDependencies": {
+    "@angular/common": "^13.0.0",
+    "@angular/core": "^13.0.0"
+  },
+  "dependencies": {
+    "tslib": "^2.3.0"
+  }
+}
