+++ projects/ngx-virtual-scroller/src/public-api.ts
@@ -1,0 +1,8 @@
+/*
+ * Public API Surface of ngx-virtual-scroller
+ */
+
+export * from './lib/ngx-virtual-scroller.module';
+export * from './lib/ngx-virtual-scroller.component';
+
+
